ABI/INFORM Trade & Industry is an industry-focused complement to ABI/INFORM
Organizations which are independent of government involvement[2] are known as non-governmental organizations or non-government organizations,[3] with NGO as an acronym.[4][5] NGOs are a subgroup of organizations founded by citizens, which include clubs and associations that provide services to their members and others. NGOs are usually nonprofit organizations, and many of them are active in humanitarianism or the social sciences. Surveys indicate that NGOs have a high degree of public trust, which can make them a useful proxy for the concerns of society and stakeholders.[6] However, NGOs can also be lobby groups for corporations, such as the World Economic Forum.[7][8][9][10] According to NGO.org (the non-governmental organizations associated with the United Nations), "[an NGO is] any non-profit, voluntary citizens' group which is organized on a local, national or international level ... Task-oriented and driven by people with a common interest, NGOs perform a variety of service and humanitarian functions, bring citizen concerns to Governments, advocate and monitor policies and encourage political participation through provision of information."[11][needs update?]

Pekka Haavisto, speaking at a podium
Pekka Haavisto, Minister for International Development of the Ministry for Foreign Affairs in Finland, at the first World NGO Day in Helsinki in 2014
A roomful of people
Europe-Georgia Institute head George Melashvili addresses the audience at the launch of the "Europe in a suitcase" project by two NGOs (the EGI and the Friedrich Naumann Foundation), which aims to increase cooperation between European politicians, journalists and representatives of the civic sector and academia with their counterparts in Georgia.[1]
Russia had about 277,000 NGOs in 2008.[12] India is estimated to have had about two million NGOs in 2009 (approximately one per 600 Indians), many more than the number of the country's primary schools and health centers.[13][14] The term "NGO" is used inconsistently; it is sometimes a synonym for a civil society organization, any association founded by citizens.[15] NGOs are known in some countries as nonprofit organizations, and political parties and trade unions are sometimes considered NGOs. NGOs are classified by orientation and level of operation; orientation refers to the type of activities an NGO undertakes. Activities may include human rights, consumer protection, environmentalism, health, or development. An NGO's level of operation indicates the scale at which an organization works: local, regional, national, or international.[16]

Types	
Activities	
Structure	
History	
Legal status	
Economic theory	
Influence on world affairs	
Criticism	
See also	
References	
Further reading	
External links	
Last edited 4 days ago by FMSky

Content is available under CC BY-SA 3.0 unless otherwise noted.
Privacy policy Terms of UseDesktop